tests/test_utils: Update behaviour of test
As part of commit #b3de85 porting to WA3, the behaviour of Enums changed
so that the atrribute to be fetched is converted to lower case before
the lookup occours. Due to the fact that 'names' is a predefined
attribute of the Enum class this clashes and will return the full list
of names instead of the individual entry. Update the test to use another
non reserved entry.